requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
